Changing released status for some Pokemon
If a pokemon is somehow added to the raid boss list but is not considered released, Kygore will choke on it. This can only happen if the raid_info.json file is manually changed. For future changes to that file, it should be ensured that all Pokemon listed there are marked as released in this file.